Move Enum doc test out of gradle integration tests
